Move App into index page to resolve issues with conflicting filename imports. Ensure all values from MapBox are serializable
